1、增加 RTC 初始化 2、上电 RTC 开始计数，打开tickless 模式 3、系统 tick 和 rtc 计数 进行比较，发现 系统tick 2100秒， RIC 才计数 2051秒
